Move container.WaitStop, AttachWithLogs and WsAttachWithLogs to daemon service in api server
